feat: final verification fixes — kwargs builder **kwargs, format_response empty choices, aioresponses API
- Add **kwargs to all 12 kwargs builder functions so dispatch layer can pass uniform parameters
- Fix format_response to return string for empty choices instead of error dict
- Fix integration tests: use payload= instead of json= for aioresponses v0.7.8
- Add test_constants.py with comprehensive constant tests (63 tests)
- Add unit tests for message_helpers, protocol, connection_pool
- Verify all 157 tests pass, syntax clean, imports working
