+++ groq_api.py
@@ -9,6 +9,7 @@ import aiohttp
 from groq import AsyncGroq, GroqError
 from if_llm.providers.base import BaseLLMProvider
 from if_llm.providers.message_helpers import build_base_messages, build_multimodal_user_message, build_text_user_message
+from if_llm.constants import CONTENT_TYPE_JSON, ImageFormat
 
 logger = logging.getLogger(__name__)
 logging.basicConfig(level=logging.DEBUG)
@@ -121,7 +122,7 @@ def prepare_groq_messages(
     
     # Add the current user message with all images if provided
     if base64_images:
-        groq_messages.append(build_multimodal_user_message(user_message, base64_images, image_format="openai"))
+        groq_messages.append(build_multimodal_user_message(user_message, base64_images, image_format=ImageFormat.OPENAI))
         logger.debug(f"Number of images sent: {len(base64_images)}")
         for idx, base64_image in enumerate(base64_images):
             logger.debug(f"Image {idx+1} Base64 Length: {len(base64_image)}")
