+++ ollama_api.py
@@ -9,6 +9,7 @@ import logging
 from if_llm.providers.base import BaseLLMProvider
 from if_llm.providers.message_helpers import build_base_messages, build_multimodal_user_message, build_text_user_message
 from if_llm.providers.connection_pool import get_session
+from if_llm.constants import CONTENT_TYPE_JSON, ImageFormat
 
 logger = logging.getLogger(__name__)
 async def create_ollama_embedding(api_base: str, model: str, prompt: Union[str, List[str]]) -> List[float]:
@@ -33,7 +34,7 @@ async def create_ollama_embedding(api_base: str, model: str, prompt: Union[str, 
     }
     
     headers = {
-        "Content-Type": "application/json"
+        "Content-Type": CONTENT_TYPE_JSON
     }
     
     async with aiohttp.ClientSession() as session:
@@ -108,7 +109,7 @@ async def send_ollama_request(api_url, base64_images, model, system_message, use
         if tool_choice:
             data["tool_choice"] = tool_choice
 
-        ollama_headers = {"Content-Type": "application/json"}
+        ollama_headers = {"Content-Type": CONTENT_TYPE_JSON}
 
         session = await get_session()
         try:
@@ -155,7 +156,7 @@ def prepare_ollama_messages(system_message, user_message, messages, base64_image
     ollama_messages = build_base_messages(system_message, messages)
 
     if base64_images:
-        ollama_messages.append(build_multimodal_user_message(user_message, base64_images, image_format="ollama"))
+        ollama_messages.append(build_multimodal_user_message(user_message, base64_images, image_format=ImageFormat.OLLAMA))
     else:
         ollama_messages.append(build_text_user_message(user_message))
