+++ omost.py
@@ -1,8 +1,9 @@
 import sys
 import os
-from typing import Dict, Any, List
-import json
-import asyncio
+import logging
+from typing import Dict, Any
+
+logger = logging.getLogger(__name__)
 
 # Add the current directory to sys.path
 current_dir = os.path.dirname(os.path.abspath(__file__))
@@ -14,7 +15,7 @@ lib_omost_dir = os.path.join(current_dir, 'lib_omost')
 if lib_omost_dir not in sys.path:
     sys.path.append(lib_omost_dir)
 
-from lib_omost.canvas import Canvas as OmostCanvas, OmostCanvasCondition, system_prompt
+from lib_omost.canvas import Canvas as OmostCanvas, system_prompt
 
 class OmostTool:
     def __init__(self, name, description, system_prompt):
@@ -22,16 +23,14 @@ class OmostTool:
         self.description = description
         self.system_prompt = system_prompt
         self.output_type = "canvas_conditioning"
-        #print(f"OmostTool initialized with system prompt: {self.system_prompt}")
+
 
     async def execute(self, args) -> Dict[str, Any]:
-        #print("DEBUG(omost.py): Entered OmostTool.execute with args:", args)
-        #print(f"OmostTool execute method called with args: {args}")
+
         prompt = args.get('input', '')
         llm_response = args.get('llm_response', '')
         
-        #print(f"Prompt: {prompt}")
-        #print(f"LLM Response: {llm_response}")
+
         
         try:
             canvas = OmostCanvas.from_bot_response(llm_response)
@@ -53,14 +52,14 @@ class OmostTool:
                 "llm_response": llm_response
             }
         except Exception as e:
-            print(f"Error processing canvas: {str(e)}")
+            logger.error(f"Error processing canvas: {str(e)}")
             result = {
                 "error": str(e),
                 "prompt": prompt,
                 "llm_response": llm_response
             }
         
-        #print(f"OmostTool execute method returning: {result}")
+
         return result
 
 async def omost_function(args: Dict[str, Any]) -> Dict[str, Any]:
