+++ tests/unit/test_connection_pool.py
@@ -1,5 +1,5 @@
 import pytest
-from if_llm.providers.connection_pool import get_session, close_pool
+from if_llm.providers.connection_pool import get_session, close_pool, session
 
 
 @pytest.fixture(autouse=True)
@@ -11,9 +11,9 @@ async def cleanup_pool():
 
 @pytest.mark.asyncio
 async def test_get_session_creates_new():
-    session = await get_session()
-    assert session is not None
-    assert not session.closed
+    session_obj = await get_session()
+    assert session_obj is not None
+    assert not session_obj.closed
 
 
 @pytest.mark.asyncio
@@ -29,3 +29,26 @@ async def test_close_pool_resets():
     await close_pool()
     s2 = await get_session()
     assert s1 is not s2
+
+
+@pytest.mark.asyncio
+async def test_closed_session_creates_new():
+    s1 = await get_session()
+    await s1.close()
+    assert s1.closed
+    s2 = await get_session()
+    assert not s2.closed
+
+
+@pytest.mark.asyncio
+async def test_session_context_manager():
+    async with session() as s:
+        assert s is not None
+        assert not s.closed
+
+
+@pytest.mark.asyncio
+async def test_session_context_manager_reuses():
+    async with session() as s1:
+        async with session() as s2:
+            assert s1 is s2
