+++ tests/unit/test_message_helpers.py
@@ -65,6 +65,48 @@ class TestBuildMultimodalUserMessage:
             "images": ["abc123"],
         }
 
+    def test_anthropic_format(self):
+        result = build_multimodal_user_message(
+            "Describe this", ["abc123"], image_format="anthropic"
+        )
+        assert result["role"] == "user"
+        assert isinstance(result["content"], list)
+        assert len(result["content"]) == 2
+        assert result["content"][0]["type"] == "text"
+        assert result["content"][0]["text"] == "Describe this"
+        assert result["content"][1]["type"] == "image"
+        assert result["content"][1]["source"]["type"] == "base64"
+        assert result["content"][1]["source"]["data"] == "abc123"
+
+    def test_gemini_format(self):
+        result = build_multimodal_user_message(
+            "Describe this", ["abc123"], image_format="gemini"
+        )
+        assert result["role"] == "user"
+        assert "parts" in result
+        assert len(result["parts"]) == 2
+        assert result["parts"][0]["text"] == "Describe this"
+        assert result["parts"][1]["inline_data"]["mime_type"] == "image/jpeg"
+        assert result["parts"][1]["inline_data"]["data"] == "abc123"
+
+    def test_multiple_images_openai(self):
+        result = build_multimodal_user_message(
+            "Compare these", ["img1", "img2"], image_format="openai"
+        )
+        assert len(result["content"]) == 3
+
+    def test_multiple_images_anthropic(self):
+        result = build_multimodal_user_message(
+            "Compare these", ["img1", "img2"], image_format="anthropic"
+        )
+        assert len(result["content"]) == 3
+
+    def test_multiple_images_gemini(self):
+        result = build_multimodal_user_message(
+            "Compare these", ["img1", "img2"], image_format="gemini"
+        )
+        assert len(result["parts"]) == 3
+
     def test_invalid_format(self):
         with pytest.raises(ValueError, match="Unknown image_format"):
             build_multimodal_user_message("test", ["img"], image_format="invalid")
